+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 2.0.2
 ```
     * Pass original total and currency as custom fields. When DMN come, check them.
+    * Pass currency in Refund request.
 ```
 
 # 2.0.1